v2_migration/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
hiredis==2.2.3

# Image processing
Pillow==10.1.0
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# CLI tools
questionary==2.0.1
rich==13.7.0

# Future: Embeddings (commented out for now)
# qdrant-client==1.7.0
# sentence-transformers==2.2.2


